+++ modules/@apostrophecms/image/ui/apos/components/AposMediaManagerDisplay.vue
@@ -2,7 +2,7 @@
   <div class="apos-media-manager-display">
     <div class="apos-media-manager-display__grid">
       <AposMediaUploader
-        v-if="canEdit"
+        v-if="moduleOptions.canCreate"
         :disabled="maxReached"
         :action="moduleOptions.action"
         :accept="accept"
@@ -83,10 +83,6 @@ export default {
     event: 'change'
   },
   props: {
-    canEdit: {
-      type: Boolean,
-      default: false
-    },
     maxReached: {
       type: Boolean,
       default: false