+++ CHANGELOG.md
@@ -8,13 +8,17 @@
 * Add testing for the `float` schema field query builder.
 * Add testing for the `integer` schema field query builder.
 * Add support for link HTML attributes in the rich text widget via configurable fields `linkFields`, extendable on a project level (same as it's done for `fields`). Add an `htmlAttribute` property to the standard fields that map directly to an HTML attribute, except `href` (see special case below), and set it accordingly, even if it is the same as the field name. Setting `htmlAttribute: 'href'` is not allowed and will throw a schema validation exception (on application boot).
-* Adds support in can and criteria methods for `create` and `delete`.
+* Adds support in `can` and `criteria` methods for `create` and `delete`.
+* Changes support for image upload from `canEdit` to `canCreate`.
+* The media manager is compatible with per-doc permissions granted via the `@apostrophecms-pro/advanced-permission` module.
 
 ### Fixes
 
 * Fix the `launder` and `finalize` methods of the `float` schema field query builder.
 * Fix the `launder` and `finalize` methods of the `integer` schema field query builder.
-
+* A user who has permission to `publish` a particular page should always be allowed to insert it into the
+published version of the site even if they could not otherwise insert a child of the published
+parent.
 
 ## 3.61.1 (2023-01-08)